Remove obsolute simulation_base class
