+++ NuRadioMC/simulation/simulation.py
@@ -37,7 +37,6 @@ import yaml
 import os
 import collections
 from NuRadioMC.utilities.Veff import remove_duplicate_triggers
-import NuRadioMC.simulation.simulation_base
 import NuRadioMC.simulation.simulation_input_output
 import NuRadioMC.simulation.channel_efield_simulator
 import NuRadioMC.simulation.shower_simulator
@@ -45,26 +44,249 @@ import NuRadioMC.simulation.station_simulator
 import NuRadioMC.simulation.hardware_response_simulator
 import NuRadioMC.simulation.output_writer_hdf5
 import NuRadioMC.simulation.output_writer_nur
-STATUS = 31
 
 # logging.basicConfig(format='%(asctime)s %(levelname)s:%(name)s:%(message)s')
 
 
 
-logging.setLoggerClass(NuRadioMC.simulation.simulation_base.NuRadioMCLogger)
+
+STATUS = 31
+class NuRadioMCLogger(logging.Logger):
+
+    def status(self, msg, *args, **kwargs):
+        if self.isEnabledFor(STATUS):
+            self._log(STATUS, msg, args, **kwargs)
+
+logging.setLoggerClass(NuRadioMCLogger)
 logging.addLevelName(STATUS, 'STATUS')
 logger = logging.getLogger("NuRadioMC")
-assert isinstance(logger, NuRadioMC.simulation.simulation_base.NuRadioMCLogger)
-# formatter = logging.Formatter('%(asctime)s %(levelname)s:%(name)s:%(message)s')
-# ch = logging.StreamHandler()
-# ch.setFormatter(formatter)
-# logger.addHandler(ch)
+print(logger.__class__)
+assert isinstance(logger, NuRadioMCLogger)
+
+
+def pretty_time_delta(seconds):
+    seconds = int(seconds)
+    days, seconds = divmod(seconds, 86400)
+    hours, seconds = divmod(seconds, 3600)
+    minutes, seconds = divmod(seconds, 60)
+    if days > 0:
+        return '%dd%dh%dm%ds' % (days, hours, minutes, seconds)
+    elif hours > 0:
+        return '%dh%dm%ds' % (hours, minutes, seconds)
+    elif minutes > 0:
+        return '%dm%ds' % (minutes, seconds)
+    else:
+        return '%ds' % (seconds,)
+
+
+def merge_config(user, default):
+    if isinstance(user, dict) and isinstance(default, dict):
+        for k, v in iteritems(default):
+            if k not in user:
+                user[k] = v
+            else:
+                user[k] = merge_config(user[k], v)
+    return user
+
 
 
 
 class simulation(
     NuRadioMC.simulation.simulation_input_output.simulation_input_output
 ):
+
+    def __init__(self, inputfilename,
+                 outputfilename,
+                 detectorfile,
+                 outputfilenameNuRadioReco=None,
+                 debug=False,
+                 evt_time=datetime.datetime(2018, 1, 1),
+                 config_file=None,
+                 log_level=logging.WARNING,
+                 default_detector_station=None,
+                 default_detector_channel=None,
+                 file_overwrite=False,
+                 write_detector=True,
+                 event_list=None,
+                 log_level_propagation=logging.WARNING,
+                 ice_model=None,
+                 **kwargs):
+        """
+        initialize the NuRadioMC end-to-end simulation
+
+        Parameters
+        ----------
+        inputfilename: string, or pair
+            the path to the hdf5 file containing the list of neutrino events
+            alternatively, the data and attributes dictionary can be passed directly to the method
+        outputfilename: string
+            specify hdf5 output filename.
+        detectorfile: string
+            path to the json file containing the detector description
+        station_id: int
+            the station id for which the simulation is performed. Must match a station
+            deself._fined in the detector description
+        outputfilenameNuRadioReco: string or None
+            outputfilename of NuRadioReco detector sim file, this file contains all
+            waveforms of the triggered events
+            default: None, i.e., no output file will be written which is useful for
+            effective volume calculations
+        debug: bool
+            True activates debug mode, default False
+        evt_time: datetime object
+            the time of the events, default 1/1/2018
+        config_file: string
+            path to config file
+        log_level: logging.LEVEL
+            the log level
+        default_detector_station: int or None
+            DEPRECATED: Define reference stations in the detector JSON file instead
+        default_detector_channel: int or None
+            DEPRECATED: Define reference channels in the detector JSON file instead
+        file_overwrite: bool
+            True allows overwriting of existing files, default False
+        write_detector: bool
+            If true, the detector description is written into the .nur files along with the events
+            default True
+        event_list: None or list of ints
+            if provided, only the event listed in this list are being simulated
+        log_level_propagation: logging.LEVEL
+            the log level of the propagation module
+        ice_model: medium object (default None)
+            allows to specify a custom ice model. This model is used if the config file specifies the ice model as "custom".
+        """
+        logger.setLevel(log_level)
+        if 'write_mode' in kwargs.keys():
+            logger.warning('Parameter write_mode is deprecated. Define the output format in the config file instead.')
+        self._log_level = log_level
+        self._log_level_ray_propagation = log_level_propagation
+        config_file_default = os.path.join(os.path.dirname(__file__), 'config_default.yaml')
+        logger.status('reading default config from {}'.format(config_file_default))
+        with open(config_file_default, 'r') as ymlfile:
+            self._cfg = yaml.load(ymlfile, Loader=yaml.FullLoader)
+        if config_file is not None:
+            logger.status('reading local config overrides from {}'.format(config_file))
+            with open(config_file, 'r') as ymlfile:
+                local_config = yaml.load(ymlfile, Loader=yaml.FullLoader)
+                new_cfg = merge_config(local_config, self._cfg)
+                self._cfg = new_cfg
+
+        if self._cfg['seed'] is None:
+            # the config seeting None means a random seed. To have the simulation be reproducable, we generate a new
+            # random seed once and save this seed to the config setting. If the simulation is rerun, we can get
+            # the same random sequence.
+            self._cfg['seed'] = np.random.randint(0, 2 ** 32 - 1)
+
+        self._outputfilename = outputfilename
+        if os.path.exists(self._outputfilename):
+            msg = f"hdf5 output file {self._outputfilename} already exists"
+            if file_overwrite == False:
+                logger.error(msg)
+                raise FileExistsError(msg)
+            else:
+                logger.warning(msg)
+        self._detectorfile = detectorfile
+        self._n_reflections = int(self._cfg['propagation']['n_reflections'])
+        self._outputfilenameNuRadioReco = outputfilenameNuRadioReco
+        self._debug = debug
+        self._evt_time = evt_time
+        self._write_detector = write_detector
+        logger.status("setting event time to {}".format(evt_time))
+        self._event_group_list = event_list
+        self._evt = None
+        self._antenna_pattern_provider = NuRadioReco.detector.antennapattern.AntennaPatternProvider()
+
+        # initialize propagation module
+        self._prop = NuRadioMC.SignalProp.propagation.get_propagation_module(self._cfg['propagation']['module'])
+
+        if self._cfg['propagation']['ice_model'] == "custom":
+            if ice_model is None:
+                logger.error("ice model is set to 'custom' in config file but no custom ice model is provided.")
+                raise AttributeError("ice model is set to 'custom' in config file but no custom ice model is provided.")
+            self._ice = ice_model
+        else:
+            self._ice = NuRadioMC.utilities.medium.get_ice_model(self._cfg['propagation']['ice_model'])
+
+        self._mout = collections.OrderedDict()
+        self._mout_groups = collections.OrderedDict()
+        self._mout_attrs = collections.OrderedDict()
+
+        # read in detector positions
+        logger.status("Detectorfile {}".format(os.path.abspath(self._detectorfile)))
+        self._det = None
+        if default_detector_station is not None:
+            logger.warning(
+                'Deprecation warning: Passing the default detector station is deprecated. Default stations and default'
+                'channel should be specified in the detector description directly.'
+            )
+            logger.status(f"Default detector station provided (station {default_detector_station}) -> Using generic detector")
+            self._det = NuRadioReco.detector.generic_detector.GenericDetector(json_filename=self._detectorfile, default_station=default_detector_station,
+                                                 default_channel=default_detector_channel, antenna_by_depth=False)
+        else:
+            self._det = NuRadioReco.detector.detector.Detector(json_filename=self._detectorfile, antenna_by_depth=False)
+
+        self._det.update(evt_time)
+
+        self._station_ids = self._det.get_station_ids()
+        self._event_ids_counter = {}
+        for station_id in self._station_ids:
+            self._event_ids_counter[station_id] = -1  # we initialize with -1 becaue we increment the counter before we use it the first time
+
+        # print noise information
+        logger.status("running with noise {}".format(bool(self._cfg['noise'])))
+        logger.status("setting signal to zero {}".format(bool(self._cfg['signal']['zerosignal'])))
+        if bool(self._cfg['propagation']['focusing']):
+            logger.status("simulating signal amplification due to focusing of ray paths in the firn.")
+
+        # read sampling rate from config (this sampling rate will be used internally)
+        self._dt = 1. / (self._cfg['sampling_rate'] * units.GHz)
+
+        if isinstance(inputfilename, str):
+            logger.status(f"reading input from {inputfilename}")
+            self._inputfilename = inputfilename
+            self._read_input_hdf5()  # we read in the full input file into memory at the beginning to limit io to the beginning and end of the run
+        else:
+            logger.status("getting input on-the-fly")
+            self._inputfilename = "on-the-fly"
+            self._fin = inputfilename[0]
+            self._fin_attrs = inputfilename[1]
+            self._fin_stations = {}
+        # store all relevant attributes of the input file in a dictionary
+        self._generator_info = {}
+        self._particle_mode = "simulation_mode" not in self._fin_attrs or self._fin_attrs['simulation_mode'] != "emitter"
+
+        for enum_entry in genattrs:
+            if enum_entry.name in self._fin_attrs:
+                self._generator_info[enum_entry] = self._fin_attrs[enum_entry.name]
+
+        # check if the input file contains events, if not save empty output file (for book keeping) and terminate simulation
+        if len(self._fin['xx']) == 0:
+            logger.status(f"input file {self._inputfilename} is empty")
+            return
+
+
+
+        self._distance_cut_polynomial = None
+        if self._cfg['speedup']['distance_cut']:
+            coef = self._cfg['speedup']['distance_cut_coefficients']
+            self.__distance_cut_polynomial = np.polynomial.polynomial.Polynomial(coef)
+
+            def get_distance_cut(shower_energy):
+                if shower_energy <= 0:
+                    return 100 * units.m
+                return max(100 * units.m, 10 ** self.__distance_cut_polynomial(np.log10(shower_energy)))
+
+            self._get_distance_cut = get_distance_cut
+
+        # initialize parameters to measure run times
+        self._input_time = 0.0
+        self._askaryan_time = 0.0
+        self._rayTracingTime = 0.0
+        self._detSimTime = 0.0
+        self._outputTime = 0.0
+        self._weightTime = 0.0
+        self._distance_cut_time = 0.0
+
 
     def run(self):
         """
@@ -320,13 +542,13 @@ class simulation(
             ts.append(instance.run.time[instance])
         ttot = np.sum(np.array(ts))
         for i, (name, instance, kwargs) in enumerate(self._evt.iter_modules(self._station.get_id())):
-            t = NuRadioMC.simulation.simulation_base.pretty_time_delta(ts[i])
+            t = pretty_time_delta(ts[i])
             trel = 100.*ts[i] / ttot
             output_NuRadioRecoTime += f"{name}: {t} {trel:.1f}%\n"
         logger.status(output_NuRadioRecoTime)
 
         logger.status("{:d} events processed in {} = {:.2f}ms/event ({:.1f}% input, {:.1f}% ray tracing, {:.1f}% askaryan, {:.1f}% detector simulation, {:.1f}% output, {:.1f}% weights calculation)".format(self._n_showers,
-                                                                                         NuRadioMC.simulation.simulation_base.pretty_time_delta(t_total), 1.e3 * t_total / self._n_showers,
+                                                                                         pretty_time_delta(t_total), 1.e3 * t_total / self._n_showers,
                                                                                          100 * self._input_time / t_total,
                                                                                          100 * (self._rayTracingTime - self._askaryan_time) / t_total,
                                                                                          100 * self._askaryan_time / t_total,
